+++ src/signalling/peers/mod.rs
@@ -762,7 +762,7 @@ mod tests {
             endpoints::webrtc_publish_endpoint::{
                 AudioSettings, P2pMode, VideoSettings,
             },
-            member::ControlCredential,
+            member::Credential,
             refs::SrcUri,
         },
         signalling::{
@@ -908,7 +908,7 @@ mod tests {
 
         let publisher = Member::new(
             "publisher".into(),
-            ControlCredential::Plain("test".into()),
+            Credential::Plain("test".into()),
             "test".into(),
             Duration::from_secs(10),
             Duration::from_secs(10),
@@ -916,7 +916,7 @@ mod tests {
         );
         let receiver = Member::new(
             "receiver".into(),
-            ControlCredential::Plain("test".into()),
+            Credential::Plain("test".into()),
             "test".into(),
             Duration::from_secs(10),
             Duration::from_secs(10),
@@ -1013,7 +1013,7 @@ mod tests {
 
         let publisher = Member::new(
             "publisher".into(),
-            ControlCredential::Plain("test".into()),
+            Credential::Plain("test".into()),
             "test".into(),
             Duration::from_secs(10),
             Duration::from_secs(10),
@@ -1021,7 +1021,7 @@ mod tests {
         );
         let receiver = Member::new(
             "receiver".into(),
-            ControlCredential::Plain("test".into()),
+            Credential::Plain("test".into()),
             "test".into(),
             Duration::from_secs(10),
             Duration::from_secs(10),