+++ src/signalling/room/rpc_server.rs
@@ -314,7 +314,7 @@ mod test {
 
     use crate::{
         api::control::{
-            member::ControlCredential, pipeline::Pipeline, MemberSpec, RoomSpec,
+            member::Credential, pipeline::Pipeline, MemberSpec, RoomSpec,
         },
         conf::{self, Conf},
         media::peer::tests::dummy_negotiation_sub_mock,
@@ -357,7 +357,7 @@ mod test {
 
         let member1 = MemberSpec::new(
             Pipeline::new(HashMap::new()),
-            ControlCredential::Plain("w/e".into()),
+            Credential::Plain(String::from("w/e")),
             None,
             None,
             None,
@@ -398,7 +398,7 @@ mod test {
 
         let member1 = MemberSpec::new(
             Pipeline::new(HashMap::new()),
-            ControlCredential::Plain("w/e".into()),
+            Credential::Plain(String::from("w/e")),
             None,
             None,
             None,
@@ -438,7 +438,8 @@ mod test {
 
         use actix::Addr;
         use medea_client_api_proto::{
-            CloseDescription, CloseReason, Credential, MemberId, RoomId,
+            self as client_proto, CloseDescription, CloseReason, MemberId,
+            RoomId,
         };
         use mockall::predicate::eq;
         use serial_test::serial;
@@ -452,6 +453,7 @@ mod test {
                     },
                     url::CallbackUrl,
                 },
+                member::Credential,
                 RoomElement,
             },
         };
@@ -475,7 +477,7 @@ mod test {
             let id = MemberId::from("member");
             let member = RoomElement::Member {
                 spec: Pipeline::new(HashMap::new()),
-                credentials: ControlCredential::Plain("test".into()),
+                credentials: Credential::Plain(String::from("test")),
                 on_leave,
                 on_join,
                 idle_timeout: None,
@@ -526,7 +528,7 @@ mod test {
 
                 room.connection_established(
                     MemberId::from("member"),
-                    Credential::from("test"),
+                    client_proto::Credential::from("test"),
                     Box::new(MockRpcConnection::new()),
                 )
                 .await
@@ -558,14 +560,14 @@ mod test {
                     .return_once(|_, _| Box::pin(future::ready(())));
                 room.connection_established(
                     MemberId::from("member"),
-                    Credential::from("test"),
+                    client_proto::Credential::from("test"),
                     Box::new(rpc_connection),
                 )
                 .await
                 .unwrap();
                 room.connection_established(
                     MemberId::from("member"),
-                    Credential::from("test"),
+                    client_proto::Credential::from("test"),
                     Box::new(MockRpcConnection::new()),
                 )
                 .await
@@ -583,7 +585,7 @@ mod test {
 
                 room.connection_established(
                     MemberId::from("member"),
-                    Credential::from("test"),
+                    client_proto::Credential::from("test"),
                     Box::new(MockRpcConnection::new()),
                 )
                 .await
@@ -605,7 +607,7 @@ mod test {
 
                 room.connection_established(
                     MemberId::from("member"),
-                    Credential::from("test"),
+                    client_proto::Credential::from("test"),
                     Box::new(MockRpcConnection::new()),
                 )
                 .await
@@ -624,7 +626,7 @@ mod test {
 
                 room.connection_established(
                     MemberId::from("member"),
-                    Credential::from("test"),
+                    client_proto::Credential::from("test"),
                     Box::new(MockRpcConnection::new()),
                 )
                 .await