+++ tests/e2e/signalling/mod.rs
@@ -89,7 +89,9 @@ pub struct TestMember {
     /// Whether to handle negotiation in [`TestMember`].
     auto_negotiation: bool,
 
-    ignore_room_management: bool,
+    /// Whether to handle room management ([`Event::RoomJoined`] /
+    /// [`Event::RoomLeft`]) in [`TestMember`].
+    auto_room_management: bool,
 }
 
 pub fn parse_join_room_url(url: &str) -> (Url, RoomId, MemberId, Credential) {
@@ -155,7 +157,7 @@ impl TestMember {
         on_connection_event: Option<ConnectionEventHandler>,
         deadline: Option<Duration>,
         auto_negotiation: bool,
-        ignore_room_management: bool,
+        auto_room_management: bool,
     ) -> Addr<Self> {
         let (url, room_id, member_id, token) = parse_join_room_url(url);
         let (_, framed) =
@@ -176,7 +178,7 @@ impl TestMember {
                 on_message,
                 on_connection_event,
                 auto_negotiation,
-                ignore_room_management,
+                auto_room_management,
             };
             this.authorize(member_id, token);
 
@@ -309,7 +311,7 @@ impl StreamHandler<Result<Frame, WsProtocolError>> for TestMember {
                 ServerMsg::Ping(id) => self.send_pong(id),
                 ServerMsg::Event { room_id, event } => {
                     assert_eq!(self.room_id, room_id);
-                    if self.ignore_room_management {
+                    if self.auto_room_management {
                         if matches!(
                             event,
                             Event::RoomJoined { .. } | Event::RoomLeft { .. }